fix: added PointerBarrier support (requires XInput 2.3) for better out-of-bounds detection
this commit attempts to fix the elusive "top-panel" clicks that get lost when an override redirect panel is opened such as when opening the calendar when clicking on the time & date or opening the "control panel ish" overlay when clicking on the top-panel-power button area. also, the "primary" button support has been greatly enhanced in an attempt to prevent "drawing" with the "primary" mouse button press.